// src/app/generated/blueprint.api/model/competency.ts
/*
 Copyright 2026 Carnegie Mellon University. All Rights Reserved.
 Released under a MIT (SEI)-style license. See LICENSE.md in the
 project root for license information.
*/

/**
 * Blueprint API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Competency { 
    dateCreated?: Date;
    dateModified?: Date | null;
    createdBy?: string;
    modifiedBy?: string | null;
    id?: string;
    competencyFrameworkId?: string;
    parentId?: string | null;
    idNumber?: string | null;
    shortName?: string | null;
    description?: string | null;
    descriptionFormat?: number;
    path?: string | null;
    sortOrder?: number;
    ruleType?: string | null;
    ruleOutcome?: number;
    ruleConfig?: string | null;
    scaleValues?: string | null;
    scaleConfiguration?: string | null;
    children?: Array<Competency> | null;
    relatedIdNumbers?: Array<string> | null;
}