// src/app/generated/blueprint.api/model/mselCompetency.ts
/**
 * Blueprint API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Competency } from './competency';


export interface MselCompetency {
    id?: string;
    mselId?: string;
    competencyId?: string;
    competency?: Competency;
}