// src/app/generated/blueprint.api/model/teamCompetency.ts
/**
 * Blueprint API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Competency } from './competency';


export interface TeamCompetency {
    id?: string;
    teamId?: string;
    competencyId?: string;
    competency?: Competency;
}